+++ d2/src/main/java/com/linkedin/d2/xds/XdsClientImpl.java
@@ -50,6 +50,7 @@ import java.util.List;
 import java.util.Map;
 import java.util.Objects;
 import java.util.Set;
+import java.util.concurrent.Executors;
 import java.util.concurrent.ScheduledExecutorService;
 import java.util.concurrent.ScheduledFuture;
 import java.util.concurrent.TimeUnit;
@@ -97,7 +98,7 @@ public class XdsClientImpl extends XdsClient
   private BackoffPolicy _retryBackoffPolicy;
   @VisibleForTesting
   AdsStream _adsStream;
-  private boolean _shutdown;
+  private boolean _isXdsStreamShutdown;
   @VisibleForTesting
   ScheduledFuture<?> _retryRpcStreamFuture;
   private ScheduledFuture<?> _readyTimeoutFuture;
@@ -175,7 +176,7 @@ public class XdsClientImpl extends XdsClient
   @Override
   public void watchXdsResource(String resourceName, ResourceWatcher watcher)
   {
-    _executorService.execute(() ->
+    checkShutdownAndExecute(() ->
     {
       ResourceType originalType = watcher.getType();
       Map<String, ResourceSubscriber> resourceSubscriberMap = getResourceSubscriberMap(originalType);
@@ -214,7 +215,7 @@ public class XdsClientImpl extends XdsClient
   @Override
   public void watchAllXdsResources(WildcardResourceWatcher watcher)
   {
-    _executorService.execute(() ->
+    checkShutdownAndExecute(() ->
     {
       ResourceType originalType = watcher.getType();
       WildcardResourceSubscriber subscriber = getWildcardResourceSubscriber(originalType);
@@ -243,7 +244,7 @@ public class XdsClientImpl extends XdsClient
   @Override
   public void startRpcStream()
   {
-    _executorService.execute(() ->
+    checkShutdownAndExecute(() ->
     {
       if (!isInBackoff())
       {
@@ -269,7 +270,7 @@ public class XdsClientImpl extends XdsClient
   @VisibleForTesting
   void startRpcStreamLocal()
   {
-    if (_shutdown)
+    if (_isXdsStreamShutdown)
     {
       _log.warn("RPC stream cannot be started after shutdown!");
       return;
@@ -284,7 +285,7 @@ public class XdsClientImpl extends XdsClient
         AggregatedDiscoveryServiceGrpc.newStub(_managedChannel);
     AdsStream stream = new AdsStream(stub);
     _adsStream = stream;
-    _readyTimeoutFuture = _executorService.schedule(() ->
+    _readyTimeoutFuture = checkShutdownAndSchedule(() ->
     {
       // There is a race condition where the task can be executed right as it's being cancelled. This checks whether
       // the current state is still pointing to the right stream, and whether it is ready before notifying of an error.
@@ -305,15 +306,38 @@ public class XdsClientImpl extends XdsClient
   @Override
   public void shutdown()
   {
-    _executorService.execute(() ->
+    // create a new executor instead of using the _executorService because this shutdown should run
+    // regardless of whether _executorService is available.
+    Executors.newSingleThreadExecutor().execute(() ->
     {
-      _shutdown = true;
+      _isXdsStreamShutdown = true;
       _log.info("Shutting down");
       if (_adsStream != null)
       {
         _adsStream.close(Status.CANCELLED.withDescription("shutdown").asException());
       }
     });
+  }
+
+  private ScheduledFuture<?> checkShutdownAndSchedule(Runnable runnable, long delay, TimeUnit unit) {
+    if (_executorService.isShutdown())
+    {
+      _log.warn("Attempting to schedule a task after _executorService was shutdown, will do nothing");
+      return null;
+    }
+
+    return _executorService.schedule(runnable, delay, unit);
+  }
+
+  private void checkShutdownAndExecute(Runnable runnable)
+  {
+    if (_executorService.isShutdown())
+    {
+      _log.warn("Attempting to execute a task after _executorService was shutdown, will do nothing");
+      return;
+    }
+
+    _executorService.execute(runnable);
   }
 
   @Override
@@ -1093,7 +1117,7 @@ public class XdsClientImpl extends XdsClient
       return;
     }
     _adsStream = testStream;
-    _retryRpcStreamFuture = _executorService.schedule(new RpcRetryTask(), 0, TimeUnit.NANOSECONDS);
+    _retryRpcStreamFuture = checkShutdownAndSchedule(new RpcRetryTask(), 0, TimeUnit.NANOSECONDS);
   }
 
   // Return true if the client should subscribe to URI glob collection for the given resource type.
@@ -1329,13 +1353,13 @@ public class XdsClientImpl extends XdsClient
             @Override
             public void beforeStart(ClientCallStreamObserver<DeltaDiscoveryRequest> requestStream)
             {
-              requestStream.setOnReadyHandler(() -> _executorService.execute(XdsClientImpl.this::readyHandler));
+              requestStream.setOnReadyHandler(() -> checkShutdownAndExecute((XdsClientImpl.this::readyHandler)));
             }
 
             @Override
             public void onNext(DeltaDiscoveryResponse response)
             {
-              _executorService.execute(() ->
+              checkShutdownAndExecute(() ->
               {
                 if (_closed)
                 {
@@ -1365,13 +1389,13 @@ public class XdsClientImpl extends XdsClient
             @Override
             public void onError(Throwable t)
             {
-              _executorService.execute(() -> handleRpcError(t));
+              checkShutdownAndExecute((() -> handleRpcError(t)));
             }
 
             @Override
             public void onCompleted()
             {
-              _executorService.execute(() -> handleRpcCompleted());
+              checkShutdownAndExecute(() -> handleRpcCompleted());
             }
           };
       _requestWriter = _stub.withWaitForReady().deltaAggregatedResources(responseReader);
@@ -1447,7 +1471,7 @@ public class XdsClientImpl extends XdsClient
         delayNanos = _retryBackoffPolicy.nextBackoffNanos();
       }
       _log.info("Retry ADS stream in {} ns", delayNanos);
-      _retryRpcStreamFuture = _executorService.schedule(new RpcRetryTask(), delayNanos, TimeUnit.NANOSECONDS);
+      _retryRpcStreamFuture = checkShutdownAndSchedule(new RpcRetryTask(), delayNanos, TimeUnit.NANOSECONDS);
     }
 
     private void close(Exception error)